+++ YoutubeFeeder/Features/FeedCache/FeedCacheCoordinator+RemoteSearch.swift
@@ -395,8 +395,3 @@ extension FeedCacheCoordinator {
         return metadata
     }
 }
-
-struct RemoteSearchTaskKey: Hashable {
-    let keyword: String
-    let limit: Int
-}
