+++ YoutubeFeeder/Features/FeedCache/FeedCacheCoordinator+RemoteSearchTaskKey.swift
@@ -1,0 +1,4 @@
+struct RemoteSearchTaskKey: Hashable {
+    let keyword: String
+    let limit: Int
+}
