+++ YoutubeFeeder/Shared/AppLogic+View.swift
@@ -1,0 +1,7 @@
+import SwiftUI
+
+extension View {
+    func bindRefreshCommand(_ action: (() async -> Void)?) -> some View {
+        modifier(RefreshCommandRegistrationModifier(action: action))
+    }
+}
